[ot] hw/opentitan: ot_otp: merge OTP implementation from EarlGrey and Darjeeling
- merge and adapt common code into OtOTPEngine
- keep only OTP top-specific implementation in ot_otp_eg.c and ot_otp_dj.c,
  which implement the OtOTPImplIf interface (specialization for Tops)
- regenerate OTP partition descriptions and key seed definition with otptool.py
