+++ hw/opentitan/Kconfig
@@ -124,12 +124,23 @@ config OT_OTP_BE_IF
     bool
 
 config OT_OTP_DJ
+    select OT_OTP_ENGINE
     select OT_OTP_IF
+    select OT_OTP_IMPL_IF
     select OT_PRESENT
     bool
 
 config OT_OTP_EG
+    select OT_OTP_ENGINE
     select OT_OTP_IF
+    select OT_OTP_IMPL_IF
+    select OT_PRESENT
+    bool
+
+config OT_OTP_ENGINE
+    select OT_EDN
+    select OT_OTP_IF
+    select OT_OTP_IMPL_IF
     select OT_PRESENT
     bool
