+++ include/hw/opentitan/ot_lc_ctrl.h
@@ -29,6 +29,7 @@
 #define HW_OPENTITAN_OT_LC_CTRL_H
 
 #include "qom/object.h"
+#include "hw/sysbus.h"
 
 #define TYPE_OT_LC_CTRL "ot-lc_ctrl"
 OBJECT_DECLARE_TYPE(OtLcCtrlState, OtLcCtrlClass, OT_LC_CTRL)
